[MOV] move add_product_option to html_builder
