+++ addons/html_builder/static/src/plugins/add_product_option.js
