--- addons/mass_mailing/static/src/builder/options/add_product_option.js
@@ -1,9 +1,0 @@
-import { BaseOptionComponent } from "@html_builder/core/utils";
-
-export class AddProductOption extends BaseOptionComponent {
-    static template = "mass_mailing.AddProductOption";
-    static props = {
-        applyTo: String,
-        productSelector: String,
-    };
-}
