--- addons/mass_mailing/static/src/builder/options/add_product_option.xml
@@ -1,16 +1,0 @@
-<templates xml:space="preserve">
-
-    <t t-name="mass_mailing.AddProductOption">
-        <BuilderRow label.translate="Product">
-            <BuilderButton
-                action="'addItem'"
-                actionParam="`${props.productSelector}:last-of-type`"
-                preview="false"
-                className="'o_we_bg_brand_primary'"
-                applyTo="props.applyTo">
-                Add Product
-            </BuilderButton>
-        </BuilderRow>
-    </t>
-
-</templates>
